==> TAREAS/tarea1_future_sales/notebooks/notebooks_anotherform/02_ventas_mensual.ipynb <==
import pandas as pd
import numpy as np
from itertools import product
import os

os.getcwd()

icategories = pd.read_csv("../data/archive/item_categories_en.csv")
items = pd.read_csv("../data/archive/items_en.csv")
shops = pd.read_csv("../data/archive/shops_en.csv")
train = pd.read_csv("../data/archive/sales_train.csv")
test = pd.read_csv("../data/archive/test.csv")
sample = pd.read_csv("../data/archive/sample_submission.csv")

train["date"] = pd.to_datetime(train["date"], format="%d.%m.%Y")
train["month"] = train["date_block_num"].astype(int)


#Limpieza de outliers

train = train[(train["item_price"] > 0) & (train["item_price"] < 100000)]
print(train.shape)


monthly = (
    train
    .groupby(["date_block_num", "shop_id", "item_id"], as_index=False)
    .agg(item_cnt_month=("item_cnt_day", "sum"))
)

print(monthly.shape)
monthly.head()


monthly["item_cnt_month"] = monthly["item_cnt_month"].clip(0, 20)
monthly["item_cnt_month"].describe()


index_cols = ["date_block_num", "shop_id", "item_id"]

grids = []
for month in monthly["date_block_num"].unique():
    cur = train[train["date_block_num"] == month]
    shops_m = cur["shop_id"].unique()
    items_m = cur["item_id"].unique()
    grids.append(
        pd.DataFrame(list(product([month], shops_m, items_m)), columns=index_cols)
    )

grid = pd.concat(grids, ignore_index=True)
print("grid:", grid.shape)
grid.head()


data = grid.merge(monthly, on=index_cols, how="left")
data["item_cnt_month"] = data["item_cnt_month"].fillna(0).astype(np.float32)

print(data.shape)
data.head()


data = data.merge(items[["item_id", "item_category_id"]], on="item_id", how="left")

# sanity check:
data.isna().sum().head(10)


data.to_csv("../data/new/monthly_dataset.csv", index=False)
print("Guardado")
