==> TAREAS/tarea1_future_sales/notebooks/notebooks_anotherform/03_Feature_Engineering.ipynb <==
import pandas as pd
import numpy as np


data = pd.read_csv("../data/new/monthly_dataset.csv")

#Ordenamos porque usaremos lags
data = data.sort_values(
    ["shop_id", "item_id", "date_block_num"]
).reset_index(drop=True)

for lag in [1, 2, 3]:
    data[f"lag_{lag}"] = (
        data
        .groupby(["shop_id", "item_id"])["item_cnt_month"] #Separamos las series de cada combinación tienda-item
        .shift(lag) #por shit() es que ordenamos antes la data, ya que shit() asume que las series están en el orden correcto
    )


#Convertimos todas las variables a promedios mensuales y les aplicamos el lag t-1
#Promedio histórico mensual del item
data["item_avg"] = (
    data
    .groupby("item_id")["item_cnt_month"]
    .transform("mean")
)


#Promedio histórico del item con lag t - 1
data["item_avg_lag_1"] = (
    data
    .groupby("item_id")["item_avg"]
    .shift(1)
)


#Promedio histórico mensual de ventas por tienda
shop_avg = (
    data
    .groupby(["date_block_num", "shop_id"], as_index=False)["item_cnt_month"]
    .mean()
    .rename(columns={"item_cnt_month": "shop_month_avg"})
)

data = data.merge(shop_avg, on=["date_block_num", "shop_id"], how="left")


#Promedio histórico mensual de ventas por tienda lag t-1
data["shop_month_avg_lag_1"] = (
    data
    .groupby("shop_id")["shop_month_avg"]
    .shift(1)
)


#Promedio histórico mensual por categoría
cat_avg = (
    data
    .groupby(["date_block_num", "item_category_id_x"], as_index=False)["item_cnt_month"]
    .mean()
    .rename(columns={"item_cnt_month": "cat_month_avg"})
)

data = data.merge(cat_avg, on=["date_block_num", "item_category_id_x"], how="left")


#Promedio histórico mensual por categoría t-1
data["cat_month_avg_lag_1"] = (
    data
    .groupby("item_category_id_x")["cat_month_avg"]
    .shift(1)
)


#Creamos un contador 0-11 para los meses, con esto vemos la relación que tienen los meses entre si (estacionalidad)

data["month_of_year"] = data["date_block_num"] % 12


#Definimos variable dependiente e independientes

FEATURES = [
    "lag_1", "lag_2", "lag_3",
    "item_avg_lag_1",
    "shop_month_avg_lag_1",
    "cat_month_avg_lag_1",
    "month_of_year"
]

TARGET = "item_cnt_month"


data.to_csv("../data/new/feature_dataset.csv", index=False)
print("guardado")
