==> TAREAS/tarea1_future_sales/notebooks/notebooks_anotherform/04_baseline.ipynb <==
import pandas as pd
import numpy as np
from sklearn.metrics import mean_squared_error


data = pd.read_csv("../data/new/monthly_dataset.csv")
print(data.shape)
data.head()


data = data.sort_values(["shop_id", "item_id", "date_block_num"])


data["lag_1"] = (
    data
    .groupby(["shop_id", "item_id"])["item_cnt_month"]
    .shift(1)
)


train_data = data[data["date_block_num"] < 33]
valid_data = data[data["date_block_num"] == 33]


y_true = valid_data["item_cnt_month"]
y_pred = valid_data["lag_1"].fillna(0)


mse = mean_squared_error(y_true, y_pred)
rmse = np.sqrt(mse)
rmse


#Chequeo rápido

(y_pred == 0).mean()
(y_pred - y_true).describe()
