==> TAREAS/tarea1_future_sales/notebooks/notebooks_anotherform/05_Modelos1_Lightgbm.ipynb <==
import pandas as pd
import numpy as np

from sklearn.metrics import mean_squared_error
import lightgbm as lgb


data = pd.read_csv("../data/new/feature_dataset.csv")

data = data.sort_values(
    ["shop_id", "item_id", "date_block_num"]
).reset_index(drop=True)

print(data.shape)
data.head()


FEATURES = [
    "lag_1", "lag_2", "lag_3",
    "item_avg_lag_1",
    "shop_month_avg_lag_1",
    "cat_month_avg_lag_1",
    "month_of_year"
]

TARGET = "item_cnt_month"


#dividimos en set de train y de validación

train_df = data[data["date_block_num"] < 33]
valid_df = data[data["date_block_num"] == 33]

#separamos las variables independientes y la dependiente
X_train = train_df[FEATURES]
y_train = train_df[TARGET]

X_valid = valid_df[FEATURES]
y_valid = valid_df[TARGET]

print(X_train.shape, X_valid.shape)


model = lgb.LGBMRegressor(
    n_estimators=300,
    learning_rate=0.05,
    num_leaves=31,
    subsample=0.8,
    colsample_bytree=0.8,
    random_state=42
)


model.fit(X_train, y_train)


y_pred = model.predict(X_valid)
y_pred = np.clip(y_pred, 0, 20)


rmse_lgbm = np.sqrt(mean_squared_error(y_valid, y_pred))
rmse_lgbm


importances = pd.Series(
    model.feature_importances_,
    index=FEATURES
).sort_values(ascending=False)

importances


import joblib

joblib.dump(model, "../modelos/lightgbm_model.joblib")
print("Modelo guardado")
