==> TAREAS/tarea1_future_sales/notebooks/notebooks_anotherform/06_Entregable.ipynb <==
import pandas as pd
import numpy as np
import joblib


model = joblib.load("../modelos/lightgbm_model.joblib")
print(" modelo cargado")


items = pd.read_csv("../data/archive/items_en.csv")
test = pd.read_csv("../data/archive/test.csv")
monthly = pd.read_csv("../data/new/monthly_dataset.csv")

if "item_category_id" not in monthly.columns:
    monthly = monthly.merge(items[["item_id", "item_category_id"]], on="item_id", how="left")


last_month = monthly["date_block_num"].max()
pred_month = last_month + 1

last_month, pred_month


pred_base = test[["shop_id", "item_id"]].copy()
pred_base["date_block_num"] = pred_month


pred_base = pred_base.merge(items[["item_id", "item_category_id"]], on="item_id", how="left")
pred_base.head()


FEATURES = [
    "lag_1", "lag_2", "lag_3",
    "item_avg_lag_1",
    "shop_month_avg_lag_1",
    "cat_month_avg_lag_1",
    "month_of_year"
]


hist = monthly[["date_block_num", "shop_id", "item_id", "item_cnt_month"]].copy()


lag1 = hist[hist["date_block_num"] == pred_month - 1].rename(columns={"item_cnt_month": "lag_1"})
lag1 = lag1[["shop_id", "item_id", "lag_1"]]
pred_base = pred_base.merge(lag1, on=["shop_id", "item_id"], how="left")


lag2 = hist[hist["date_block_num"] == pred_month - 2].rename(columns={"item_cnt_month": "lag_2"})
lag2 = lag2[["shop_id", "item_id", "lag_2"]]
pred_base = pred_base.merge(lag2, on=["shop_id", "item_id"], how="left")


lag3 = hist[hist["date_block_num"] == pred_month - 3].rename(columns={"item_cnt_month": "lag_3"})
lag3 = lag3[["shop_id", "item_id", "lag_3"]]
pred_base = pred_base.merge(lag3, on=["shop_id", "item_id"], how="left")


for c in ["lag_1", "lag_2", "lag_3"]:
    pred_base[c] = pred_base[c].fillna(0)


item_avg = (
    hist.groupby("item_id")["item_cnt_month"].mean()
    .rename("item_avg_lag_1")
    .reset_index()
)
pred_base = pred_base.merge(item_avg, on="item_id", how="left")
pred_base["item_avg_lag_1"] = pred_base["item_avg_lag_1"].fillna(0)


shop_month_avg_last = (
    hist[hist["date_block_num"] == pred_month - 1]
    .groupby("shop_id")["item_cnt_month"]
    .mean()
    .rename("shop_month_avg_lag_1")
    .reset_index()
)

pred_base = pred_base.merge(shop_month_avg_last, on="shop_id", how="left")
pred_base["shop_month_avg_lag_1"] = pred_base["shop_month_avg_lag_1"].fillna(0)


cat_month_avg_last = (
    monthly[monthly["date_block_num"] == pred_month - 1]
    .groupby("item_category_id")["item_cnt_month"]
    .mean()
    .rename("cat_month_avg_lag_1")
    .reset_index()
)
pred_base = pred_base.merge(cat_month_avg_last, on="item_category_id", how="left")
pred_base["cat_month_avg_lag_1"] = pred_base["cat_month_avg_lag_1"].fillna(0)


pred_base["month_of_year"] = pred_base["date_block_num"] % 12


FEATURES = [
    "lag_1", "lag_2", "lag_3",
    "item_avg_lag_1",
    "shop_month_avg_lag_1",
    "cat_month_avg_lag_1",
    "month_of_year"
]

X_test = pred_base[FEATURES]
pred = model.predict(X_test)
pred = np.clip(pred, 0, 20)


submission = pd.DataFrame({
    "ID": test["ID"],
    "item_cnt_month": pred
})

submission.head()


submission.to_csv("../entregables/submission.csv", index=False)
print("✅ Guardado en ./submissions/submission.csv")
